requirements.txt
Eel~=0.18.2
bottle~=0.13.4
gevent-websocket~=0.10.1
gevent~=25.9.1
pyparsing~=3.3.1
importlib_resources~=6.5.2
pyinstaller~=6.18.0
pip~=25.1.1
cffi~=2.0.0
setuptools~=80.9.0
future~=1.0.0
greenlet~=3.3.0
psutil~=7.1.3
certifi~=2026.1.4
urllib3~=2.6.3
altgraph~=0.17.5
idna~=3.11
charset-normalizer~=3.4.4
packaging~=25.0
pefile~=2024.8.26
pycparser~=2.23
platformdirs~=4.5.1
Pygments~=2.19.2
pywin32-ctypes~=0.2.3
requests~=2.32.5
traitlets~=5.14.3
pyinstaller-hooks-contrib~=2025.11
jupyter_core~=5.9.1
pandas~=2.3.3
pyperclip~=1.11.0
opencv-python~=4.12.0.88
schedule~=1.2.2
mss~=10.1.0
mysql-connector-python~=9.5.0
pwinput~=1.0.3
openai-whisper~=20250625
Jinja2~=3.1.6
filelock~=3.20.3
dnspython~=2.8.0
more-itertools~=10.8.0
pyfiglet~=1.0.4
pytz~=2025.2
pymongo~=4.16.0
matplotlib~=3.10.8
pyreadline3~=3.5.4
emoji~=2.15.0
protobuf~=6.33.5
ttkthemes~=3.3.0
torch~=2.10.0
transformers~=5.1.0
argostranslate~=1.11.0
PyQt6~=6.10.2
PyQt6_sip~=13.11.0
QtPy~=2.4.3
Flask~=3.1.2
numba~=0.63.1
colorama~=0.4.6
numpy~=2.2.6
scikit-learn~=1.8.0
seaborn~=0.13.2